fix tests.py in users app
